# LibCrypto: Add x86 specific versions of SHA1 and SHA2