--- a/Userland/Libraries/LibCrypto/Hash/SHA1.h
+++ b/Userland/Libraries/LibCrypto/Hash/SHA1.h
@@ -67,7 +67,7 @@ public:
     }
 
 private:
-    inline void transform(u8 const*);
+    inline void transform(u8 const (&data)[BlockSize]);
 
     u8 m_data_buffer[BlockSize] {};
     size_t m_data_length { 0 };
@@ -76,7 +76,6 @@ private:
     u32 m_state[5];
 
     constexpr static auto FinalBlockDataSize = BlockSize - 8;
-    constexpr static auto Rounds = 80;
 };
 
 }