CC @MarekKnapek
I revived your PR, I hope you dont mind me giving my self co-author

---

Changes since #23958:
* We now use a custom ifunc resolver, as the `sha` target seems to be silently ignored at best by all compilers