@MarekKnapek 2024-06-02
Yes, no problem, you have done lot of work on this and helped me much, I thank you for this. Now, towards AES... should be easier than this.