@Hendiadyoin1 2024-06-10
Before (Clang):
```
Benchmarking sha1...
Running benchmark for sha1 with size 16 for ~3000ms...3001ms, 8391481 ops, 14.3 MiB/s
Running benchmark for sha1 with size 1024 for ~3000ms...3001ms, 907621 ops, 240.3 MiB/s
Running benchmark for sha1 with size 16384 for ~3000ms...3001ms, 62750 ops, 324.2 MiB/s
Running benchmark for sha1 with size 262144 for ~3000ms...3001ms, 3931 ops, 327.1 MiB/s
Running benchmark for sha1 with size 1048576 for ~3000ms...3002ms, 960 ops, 319.4 MiB/s
Running benchmark for sha1 with size 16777216 for ~3000ms...3036ms, 62 ops, 326.1 MiB/s
Algorithm            Size       Min us/op  Max us/op  Avg us/op  Throughput
sha1                 16 B       1          112        1          14.3 MiB  /s
sha1                 1.0 KiB    4          57         4          240.3 MiB /s
sha1                 16.0 KiB   46         121        48         324.2 MiB /s
sha1                 256.0 KiB  732        1029       763        327.1 MiB /s
sha1                 1.0 MiB    3003       5511       3126       319.4 MiB /s
sha1                 16.0 MiB   48587      49801      48965      326.1 MiB /s
```
After:
Benchmarking sha1...
Running benchmark for sha1 with size 16 for ~3000ms...3001ms, 14939406 ops, 14.3 MiB/s
Running benchmark for sha1 with size 1024 for ~3000ms...3001ms, 4136306 ops, 965.1 MiB/s
Running benchmark for sha1 with size 16384 for ~3000ms...3001ms, 336116 ops, 1.6 GiB/s
Running benchmark for sha1 with size 262144 for ~3000ms...3001ms, 21397 ops, 1.7 GiB/s
Running benchmark for sha1 with size 1048576 for ~3000ms...3001ms, 5331 ops, 1.7 GiB/s
Running benchmark for sha1 with size 16777216 for ~3000ms...3003ms, 331 ops, 1.7 GiB/s
Algorithm            Size       Min us/op  Max us/op  Avg us/op  Throughput
sha1                 16 B       1          62         1          14.3 MiB  /s
sha1                 1.0 KiB    1          121        1          965.1 MiB /s
sha1                 16.0 KiB   9          104        9          1.6 GiB   /s
sha1                 256.0 KiB  135        310        140        1.7 GiB   /s
sha1                 1.0 MiB    536        759        563        1.7 GiB   /s
sha1                 16.0 MiB   8899       9256       9070       1.7 GiB   /s
```
So a 5,3x speedup for sha on my end