@nico 2024-07-03
@gmta, do the first few SIMD commits look good to you?

(I think this is fine for merging, but you know the SIMD bits much better.)